+++ CHANGELOG.md
@@ -16,6 +16,7 @@
  - Fix not being able to drop items with a wand active.
  - Add a hard limit to range and radius multipliers
  - Make wands.yml and spells.yml additive
+ - Restore 1.6 support. The blink spell needs some cleanup.
 
 ## 2.8.8